requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
deepgram-sdk==2.12.0
google-generativeai==0.8.5
supabase==2.3.0
gotrue<2.0.0
python-multipart

